scripts/requirements.txt
pandas==1.3.3
numpy==1.21.4
google-cloud-storage
